+++ webpack.config.js
@@ -54,5 +54,5 @@ module.exports = {
 			'block/src/label',
 			'view.js'
 		),
-	}
+	},
 };