+++ CHANGELOG.md
@@ -4,9 +4,11 @@ All notable changes to the Tax Switch for WooCommerce plugin will be documented 
 ## [1.5.0] - 2025-04-26
 ### Added
 - Gutenberg block/shortcode for showing text about the currently selected tax setting. [See this topic](https://wordpress.org/support/topic/shortcode-for-wdevs-tax-switch-label-text/)
+- Compatibility for Flatsome theme
+- Compatibility for FacetWP
 ### Updated
 - Tested WooCommerce 9.8.2
-- Possible breaking change: refactored the code base for registering multiple blocks
+- Possible breaking change: refactored the (block) code structure to support multiple blocks
 
 ## [1.4.3] - 2025-04-16
 ### Updated